services/ml/common/metrics.py
from sklearn.metrics import mean_absolute_error, r2_score


def regression_metrics(y_true, y_pred):
    return {
        "mae": float(mean_absolute_error(y_true, y_pred)),
        "r2": float(r2_score(y_true, y_pred)),
    }
