services/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
prophet==1.1.6
joblib==1.4.2
requests==2.32.3
httpx==0.28.1
openai==1.58.1
google-generativeai==0.8.3
